move keygen clap logic from keygen to clap-v3-utils
